+++ backend/app/models/user.py
@@ -6,6 +6,8 @@ UserProfile stores investment preferences and UI settings.
 Business profile fields are kept on User for simplicity.
 """
 
+from decimal import Decimal
+
 from sqlalchemy import Column, String, Boolean, DateTime, JSON, ForeignKey, Float, Text, Integer, Numeric
 from sqlalchemy.dialects.postgresql import UUID, ARRAY
 from sqlalchemy.orm import relationship, Mapped, mapped_column
@@ -190,12 +192,12 @@ class UserProfile(Base):
     target_markets: Mapped[Optional[List[str]]] = mapped_column(ARRAY(String), default=list)
 
     # Budget
-    investment_budget_min: Mapped[Optional[float]] = mapped_column(Numeric(12, 2))
-    investment_budget_max: Mapped[Optional[float]] = mapped_column(Numeric(12, 2))
+    investment_budget_min: Mapped[Optional[Decimal]] = mapped_column(Numeric(12, 2))
+    investment_budget_max: Mapped[Optional[Decimal]] = mapped_column(Numeric(12, 2))
 
     # Target Returns
-    target_cash_on_cash: Mapped[Optional[float]] = mapped_column(Numeric(5, 4))
-    target_cap_rate: Mapped[Optional[float]] = mapped_column(Numeric(5, 4))
+    target_cash_on_cash: Mapped[Optional[Decimal]] = mapped_column(Numeric(5, 4))
+    target_cap_rate: Mapped[Optional[Decimal]] = mapped_column(Numeric(5, 4))
 
     # Risk Tolerance
     risk_tolerance: Mapped[Optional[str]] = mapped_column(String(20))
